c++: static constexpr strictness [PR99456]
r11-7740 limited constexpr rejection of conversion from pointer to integer
to manifestly constant-evaluated contexts; it should instead check whether
we're in strict mode.

The comment for that commit noted that making this change regressed other
tests, which turned out to be because maybe_constant_init_1 was not being
properly strict for variables declared constexpr/constinit.

	PR c++/99456

gcc/cp/ChangeLog:

	* constexpr.cc (cxx_eval_constant_expression): Check strict
	instead of manifestly_const_eval.
	(maybe_constant_init_1): Be strict for static constexpr vars.